Move what a position is named by out of the partition
A position's path, the number it is measured at, what stopped a reading
of it and a rule that reading could not take in are not conclusions of a
partition. They are the words a reading of an input position is written
in, and the partition is one of the readers.

Which was invisible while it was the only reader. The measure that
counts a parameter's cases reads its type and nothing else, so a case
the rules refuse stays in its denominator and a row nobody can build is
asked for; putting that measure on the same reading means naming these
types outside the package that happens to draw axes with them.

A pure move. Every identifier keeps its name and its javadoc, the
imports follow, and nothing chose differently: 4954 tests, no failures.
`BoundaryDomain` goes with them, since `NumericTerm.intervals` is how a
term says where the values beside a line are found and the term is here
now.
