+++ souther-compiler/src/main/java/souther/compiler/inputs/BoundaryDomain.java
@@ -1,4 +1,4 @@
-package souther.compiler.partition;
+package souther.compiler.inputs;
 
 import souther.compiler.check.Carrier;
 import souther.compiler.numeric.Count;
