+++ souther-compiler/src/main/java/souther/compiler/inputs/TermPath.java
@@ -1,4 +1,4 @@
-package souther.compiler.partition;
+package souther.compiler.inputs;
 
 import java.util.List;
