+++ souther-compiler/src/main/java/souther/compiler/partition/BoundaryTarget.java
@@ -1,6 +1,7 @@
 package souther.compiler.partition;
 
 import souther.compiler.check.Carrier;
+import souther.compiler.inputs.NumericTerm;
 import souther.compiler.numeric.Place;
 
 /**
