+++ souther-compiler/src/main/java/souther/compiler/partition/GuardReachability.java
@@ -1,5 +1,6 @@
 package souther.compiler.partition;
 
+import souther.compiler.inputs.NumericTerm;
 import souther.compiler.numeric.NumericDomain;
 
 import java.util.LinkedHashSet;
