+++ souther-compiler/src/main/java/souther/compiler/partition/Partitions.java
@@ -13,6 +13,12 @@ import souther.compiler.check.FieldDomains;
 import souther.compiler.check.Rules;
 import souther.compiler.check.NumericMeasures;
 import souther.compiler.codegen.InvariantConstraints;
+import souther.compiler.inputs.BlockReason;
+import souther.compiler.inputs.BoundaryDomain;
+import souther.compiler.inputs.Membership;
+import souther.compiler.inputs.NumericTerm;
+import souther.compiler.inputs.TermPath;
+import souther.compiler.inputs.UnreadRule;
 import souther.compiler.numeric.Count;
 import souther.compiler.numeric.Endpoint;
 import souther.compiler.numeric.Granularity;
