+++ souther-compiler/src/main/java/souther/compiler/partition/StructuralInspection.java
@@ -1,6 +1,7 @@
 package souther.compiler.partition;
 
 import souther.compiler.check.Shape;
+import souther.compiler.inputs.BlockReason;
 import souther.compiler.types.Type;
 
 import java.util.Map;
