+++ souther-compiler/src/main/java/souther/compiler/partition/Threshold.java
@@ -1,5 +1,7 @@
 package souther.compiler.partition;
 
+import souther.compiler.inputs.NumericTerm;
+import souther.compiler.inputs.TermPath;
 import souther.compiler.numeric.Place;
 
 /**
