+++ souther-compiler/src/test/java/souther/compiler/CompileExampleBoundaryTest.java
@@ -6,7 +6,7 @@ import souther.compiler.observe.MeasurementStatus;
 import souther.compiler.query.Adequacy;
 import souther.compiler.query.Compilation;
 import souther.compiler.query.BoundaryAssessment;
-import souther.compiler.partition.TermPath;
+import souther.compiler.inputs.TermPath;
 import souther.compiler.query.PartitionEvidence;
 
 import java.util.List;
