+++ souther-compiler/src/test/java/souther/compiler/partition/AGuardsArmsAreNotItsThresholdTest.java
@@ -1,6 +1,8 @@
 package souther.compiler.partition;
 
 import org.junit.jupiter.api.Test;
+import souther.compiler.inputs.NumericTerm;
+import souther.compiler.inputs.TermPath;
 import souther.compiler.types.CoverageOrigin;
 
 import souther.compiler.ast.Hir;
