+++ souther-compiler/src/test/java/souther/compiler/partition/ALeafIsNotAnAbsenceTest.java
@@ -6,6 +6,7 @@ import souther.compiler.check.Resolve;
 import souther.compiler.check.SyntaxSymbols;
 import souther.compiler.check.Symbols;
 import souther.compiler.frontend.CstFrontend;
+import souther.compiler.inputs.BlockReason;
 import souther.compiler.types.Type;
 import souther.compiler.types.TypeKey;
 import souther.compiler.types.TypeSymbols;
