+++ souther-compiler/src/test/java/souther/compiler/partition/AnOpenPositionIsAReadingThatRanToTheEndTest.java
@@ -10,6 +10,8 @@ import souther.compiler.check.SyntaxSymbols;
 import souther.compiler.check.Symbols;
 import souther.compiler.check.TypeView;
 import souther.compiler.frontend.CstFrontend;
+import souther.compiler.inputs.BlockReason;
+import souther.compiler.inputs.TermPath;
 import souther.compiler.types.Type;
 import souther.compiler.types.TypeKey;
 import souther.compiler.types.TypeSymbols;
