+++ souther-compiler/src/test/java/souther/compiler/partition/AnOrdinalIsNeverWhatAnEnumerationIsWrittenAsTest.java
@@ -3,6 +3,7 @@ package souther.compiler.partition;
 import org.junit.jupiter.api.Test;
 
 import souther.compiler.check.Carrier;
+import souther.compiler.inputs.BoundaryDomain;
 import souther.compiler.numeric.Count;
 import souther.compiler.observe.ObservedValue;
 import souther.compiler.types.TypeKey;
