+++ souther-compiler/src/test/java/souther/compiler/partition/OneProjectionWritesTheWordADocumentReadsTest.java
@@ -12,6 +12,7 @@ import java.util.List;
 import java.util.Set;
 import java.util.TreeSet;
 
+import souther.compiler.inputs.BlockReason;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
